--- Class.java
@@ -1,10 +1,0 @@
-import java.util.*;
-class Class {
-  //Field->Type
-  public HashMap<String,String> fmap;
-  //Constructor
-  public Constructor cons;
-  //Method
-  public HashMap<String,Method> methodMap;
-}
-
