--- Constraint.java
@@ -1,8 +1,0 @@
-import java.util.*;
-class Constraint {
-  //Class name
-  public String className;
-  //constraint field-type
-  public HashMap<String,String> cmap = new HashMap<String,String>();
-}
-
