+++ src/Class.java
@@ -1,0 +1,8 @@
+import java.util.*;
+class Class {
+  //Constructor
+  public Constructor cons;
+  //Method
+  public HashMap<String,Method> method;
+}
+
