+++ src/Constraint.java
@@ -1,0 +1,7 @@
+import java.util.*;
+class Constraint {
+  public String className;
+  //field->Type
+  public HashMap<String,String> fieldType = new HashMap<String,String>();
+}
+
